==> 00_preprocessing.ipynb <==
# Preprocessing

> Prepcocess data: load histo and mask images as well as tabular clinical data from directories to generate one data frame.

#| default_exp nb_00_preprocessing

#| hide
from nbdev.showdoc import *

#| export
from pathlib import Path
import os
from PIL import Image
import matplotlib.pyplot as plt
import numpy as np

Path.ls = lambda x: list(x.iterdir())

## Images

path_x = Path("/media/dimi/TOSHIBA EXT/Dimitrij Original Images")
path_y = Path("/media/dimi/TOSHIBA EXT/Dimitrij Shape Masks")

#| export
def _get_files(p, fs, extensions=None):
    p = Path(p)
    res = [p/f for f in fs if not f.startswith('.')
           and ((not extensions) or f'.{f.split(".")[-1].lower()}' in extensions)]
    return res

def get_files(path, extensions=None, recurse=False, include=None):
    "Source: fast.ai, Jeremy Howard"
    path = Path(path)
    extensions = {e.lower() for e in extensions}
    if recurse:
        res = []
        for i,(p,d,f) in enumerate(os.walk(path)): # returns (dirpath, dirnames, filenames)
            if include is not None and i==0: d[:] = [o for o in d if o in include]
            else:                            d[:] = [o for o in d if not o.startswith('.')]
            res += _get_files(p, f, extensions)
        return res
    else:
        f = [o.name for o in os.scandir(path) if o.is_file()]
        return _get_files(path, f, extensions)

fn_x = get_files(path_x, recurse=True, extensions=[".tif"])
fn_y = get_files(path_y, recurse=True, extensions=[".tif"])
len(fn_x), len(fn_y)

## Clinical data

path_c = Path("/media/dimi/TOSHIBA EXT/Dimitrij Single Cores/allTXT")

fn_c = get_files(path_c, recurse=False, extensions=[".txt"])

#| export 
import pandas as pd

## One clinical dataframe

#|export
def get_clinical_data(p):
    "p: path to directory with files"
    dfs = []
    for fn in p:
        df = pd.read_csv(fn, sep="\t", encoding = "ISO-8859-1", engine="python")
        df = df[df.columns.drop(list(df.filter(regex='Unnamed')))]
        df.insert(0, 'TMA_ID', fn.stem)
        dfs.append(df)
    return pd.concat(dfs)

df_c = get_clinical_data(fn_c)

df_c.shape

## One images dataframe

#| export
def _get_tma_id(fn):
    """fn: pathlib path to file"""
    tma = fn.parent.name.split('_')[0]
    return tma[:-2] + tma[-2:].lstrip('0')

def get_tma_id(fns):
    """fns: list of pathlib paths"""
    return list(map(_get_tma_id, fns))

#| export
def _get_tma_spot(fn):
    """fn: pathlib path to file"""
    f = fn.stem[-1]+fn.stem[:-2]
    return f if len(f)==3 else f[0]+'0'+f[1] 

def get_tma_spot(fns):
    """fns: list of pathlib paths"""
    return list(map(_get_tma_spot, fns))

df_x = pd.DataFrame({'TMA_ID': get_tma_id(fn_x), 
                     'TMA-Spot':get_tma_spot(fn_x), 
                     'fn_x':fn_x})

df_y = pd.DataFrame({'TMA_ID': get_tma_id(fn_y), 
                     'TMA-Spot':get_tma_spot(fn_y), 
                     'fn_x':fn_y})

df_x.shape

df_y.shape

df_xy = pd.merge(df_x, df_y, on=["TMA_ID", "TMA-Spot"])

df_xy.shape

df_all = pd.merge(df_xy, df_c, on=["TMA_ID", "TMA-Spot"])

df_all.shape

## All together

path_x = Path("/media/dimi/TOSHIBA EXT/Dimitrij Original Images")
path_y = Path("/media/dimi/TOSHIBA EXT/Dimitrij Shape Masks")
path_c = Path("/media/dimi/TOSHIBA EXT/Dimitrij Single Cores/allTXT")

fn_x = get_files(path_x, recurse=True, extensions=[".tif"])
fn_y = get_files(path_y, recurse=True, extensions=[".tif"])
fn_c = get_files(path_c, recurse=False, extensions=[".txt"])

df_x = pd.DataFrame({'TMA_ID': get_tma_id(fn_x), 
                     'TMA-Spot':get_tma_spot(fn_x), 
                     'fn_x':fn_x})

df_y = pd.DataFrame({'TMA_ID': get_tma_id(fn_y), 
                     'TMA-Spot':get_tma_spot(fn_y), 
                     'fn_y':fn_y})

df_c = get_clinical_data(fn_c)

df_xy = pd.merge(df_x, df_y, on=["TMA_ID", "TMA-Spot"])
df_all = pd.merge(df_xy, df_c, on=["TMA_ID", "TMA-Spot"])

np.unique(df_all.TMA_ID)

df_all["Type"] = 'NaN'

df_all.loc[df_all.TMA_ID=='TMA1', "Type"] = "Greyzone"
df_all.loc[df_all.TMA_ID=='TMA2', "Type"] = "Greyzone"
df_all.loc[df_all.TMA_ID=='TMA3', "Type"] = "Greyzone"

df_all.loc[df_all.TMA_ID=='TMA5', "Type"] = "Cohort"
df_all.loc[df_all.TMA_ID=='TMA6', "Type"] = "Cohort"
df_all.loc[df_all.TMA_ID=='TMA7', "Type"] = "Cohort"
df_all.loc[df_all.TMA_ID=='TMA8', "Type"] = "Cohort"

df_all.loc[df_all.TMA_ID=='TMA25', "Type"] = "TAM-only"
df_all.loc[df_all.TMA_ID=='TMA26', "Type"] = "TAM-only"
df_all.loc[df_all.TMA_ID=='TMA27', "Type"] = "TAM-only"
df_all.loc[df_all.TMA_ID=='TMA28', "Type"] = "TAM-only"
df_all.loc[df_all.TMA_ID=='TMA29', "Type"] = "TAM-only"

df_all.Type.unique()

## Save data frame

df_all.to_pickle("data/df_all.pkl")

#| hide
import nbdev; nbdev.nbdev_export()
